Extract terminal switch fixture
