+++ src/ILInspector.Decompiler.Tests/SwitchRaisingTerminalContinuationSamples.cs
@@ -1,0 +1,14 @@
+namespace ILInspector.Decompiler.Tests;
+
+public enum CfgTerminalSwitchKind
+{
+    Value25 = 25,
+    Value30 = 30,
+    Value31 = 31,
+    Value32 = 32,
+    Value33 = 33,
+    Value35 = 35,
+    Value36 = 36,
+    Value37 = 37,
+    Value40 = 40,
+}
